implemented message service

+++ backend/src/main/java/com/nicolas/chatapp/dto/request/SendMessageRequestDTO.java
@@ -1,4 +1,4 @@
 package com.nicolas.chatapp.dto.request;
 
-public record SendMessageRequestDTO(Integer chatId, String content, Integer userId) {
+public record SendMessageRequestDTO(Long chatId, String content, Long userId) {
 }
